refactor: extract browser probe helpers
